pyproject.toml
[project]
name = "app-short-link"
version = "0.1.0"
description = "A FastAPI based app for short link generation and redirection using REST API."
authors = [
    {name = "Krzysztof Urbanowski",email = "christopher.urbanowski@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "==3.12.3"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "tortoise-orm (>=0.24.0,<0.25.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"

